Move peer stats to the new framework
